refactor(backend): extract API error boundary
